django_offices2/requirements.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.11.1
mysqlclient @ file:///C:/Users/jyu40/Downloads/Kursy/python_projects/django_offices2/django_offices2/mysqlclient-1.4.6-cp38-cp38-win32.whl
pytz==2020.1
sqlparse==0.3.1
